Move saved booklist views
- Move saved booklist related views to it's own file
